+++ index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/src/resources/Icon.png" />
+    <link rel="icon" type="image/svg+xml" href="/src/resources/icons/sl-icon.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Sam Ly</title>
   </head>
