+++ src/App.jsx
@@ -1,5 +1,5 @@
-import Header from "./Header";
-import MainContent from "./MainContent";
+import Header from "./components/Header";
+import MainContent from "./components/MainContent";
 import "./App.css"
 
 function App() {
