--- src/components/ViewMoreButton.jsx
@@ -1,11 +1,0 @@
-import './Buttons.css'
-
-function ViewMoreButton() {
-    return (
-        <button className='important-button view-more'>
-            Learn More
-        </button>
-    )
-}
-
-export default ViewMoreButton;
