--- src/components/VisitBlogButton.jsx
@@ -1,11 +1,0 @@
-import './Buttons.css'
-
-function VisitBlogButton() {
-    return (
-        <button className='important-button visit-blog'>
-            Vist Blog
-        </button>
-    )
-}
-
-export default VisitBlogButton;
