+++ src/resources/icons/sl-icon.png
