+++ src/resources/pictures/vsa-photo.png
